+++ src/backend/types/index.ts
@@ -4,3 +4,4 @@ export * from "./queries.types";
 export * from "./repository.types";
 export type * from "./history.types";
 export type * from "./workflow.types";
+export type * from "./workingTree.types";